fix:modify int64

+++ contrib/drivers/mysql/mysql_feature_master_slave_test.go
@@ -85,16 +85,16 @@ func Test_Master_Slave(t *testing.T) {
 		// Auto slave.
 		count, err = masterSlaveDB.Model(table).Count()
 		t.AssertNil(err)
-		t.Assert(count, 0)
+		t.Assert(count, int64(0))
 
 		// slave.
 		count, err = masterSlaveDB.Model(table).Slave().Count()
 		t.AssertNil(err)
-		t.Assert(count, 0)
+		t.Assert(count, int64(0))
 
 		// master.
 		count, err = masterSlaveDB.Model(table).Master().Count()
 		t.AssertNil(err)
-		t.Assert(count, TableSize)
+		t.Assert(count, int64(TableSize))
 	})
 }
